Move ParsedRepoFilter to the query module
