+++ internal/search/query/repo_revs_test.go
@@ -1,4 +1,4 @@
-package search
+package query
 
 import (
 	"testing"
