+++ internal/search/repos/repos.go
@@ -231,9 +231,9 @@ func (r *Resolver) Resolve(ctx context.Context, op search.RepoOptions) (_ Resolv
 
 		searchContextRepositoryRevisions = make(map[api.RepoID]RepoRevSpecs, len(scRepoRevs))
 		for _, repoRev := range scRepoRevs {
-			revSpecs := make([]search.RevisionSpecifier, 0, len(repoRev.Revs))
+			revSpecs := make([]query.RevisionSpecifier, 0, len(repoRev.Revs))
 			for _, rev := range repoRev.Revs {
-				revSpecs = append(revSpecs, search.RevisionSpecifier{RevSpec: rev})
+				revSpecs = append(revSpecs, query.RevisionSpecifier{RevSpec: rev})
 			}
 			searchContextRepositoryRevisions[repoRev.Repo.ID] = RepoRevSpecs{
 				Repo: repoRev.Repo,
@@ -299,7 +299,7 @@ func (r *Resolver) associateReposWithRevs(
 		i, repo := i, repo
 		g.Go(func() {
 			var (
-				revs      []search.RevisionSpecifier
+				revs      []query.RevisionSpecifier
 				isMissing bool
 			)
 
@@ -310,7 +310,7 @@ func (r *Resolver) associateReposWithRevs(
 			}
 
 			if len(revs) == 0 {
-				var clashingRevs []search.RevisionSpecifier
+				var clashingRevs []query.RevisionSpecifier
 				revs, clashingRevs = getRevsForMatchedRepo(repo.Name, includePatternRevs)
 
 				// if multiple specified revisions clash, report this usefully:
@@ -393,7 +393,7 @@ func (r *Resolver) normalizeRefs(ctx context.Context, repoRevSpecs []RepoRevSpec
 func (r *Resolver) normalizeRepoRefs(
 	ctx context.Context,
 	repo types.MinimalRepo,
-	revSpecs []search.RevisionSpecifier,
+	revSpecs []query.RevisionSpecifier,
 	reportMissing func(RepoRevSpecs),
 ) ([]string, error) {
 	revs := make([]string, 0, len(revSpecs))
@@ -417,7 +417,7 @@ func (r *Resolver) normalizeRepoRefs(
 				if errors.Is(err, context.DeadlineExceeded) || errors.HasType(err, &gitdomain.BadCommitError{}) {
 					return nil, err
 				}
-				reportMissing(RepoRevSpecs{Repo: repo, Revs: []search.RevisionSpecifier{rev}})
+				reportMissing(RepoRevSpecs{Repo: repo, Revs: []query.RevisionSpecifier{rev}})
 				continue
 			}
 			revs = append(revs, rev.RevSpec)
@@ -661,7 +661,7 @@ func (r *Resolver) filterRepoHasFileContent(
 							if errors.Is(err, context.DeadlineExceeded) || errors.HasType(err, &gitdomain.BadCommitError{}) {
 								return err
 							}
-							addMissing(RepoRevSpecs{Repo: repo, Revs: []search.RevisionSpecifier{{RevSpec: rev}}})
+							addMissing(RepoRevSpecs{Repo: repo, Revs: []query.RevisionSpecifier{{RevSpec: rev}}})
 							return nil
 						}
 
@@ -836,7 +836,7 @@ func computeExcludedRepos(ctx context.Context, db database.DB, op search.RepoOpt
 // archive.
 func ExactlyOneRepo(repoFilters []string) bool {
 	if len(repoFilters) == 1 {
-		repoRevs := search.ParseRepositoryRevisions(repoFilters[0])
+		repoRevs := query.ParseRepositoryRevisions(repoFilters[0])
 		repo := repoRevs.Repo
 		if strings.HasPrefix(repo, "^") && strings.HasSuffix(repo, "$") {
 			filter := strings.TrimSuffix(strings.TrimPrefix(repo, "^"), "$")
@@ -865,13 +865,13 @@ type ExcludedRepos struct {
 // an actual map, because we want regexp matches, not identity matches.
 type patternRevspec struct {
 	includePattern *regexp.Regexp
-	revs           []search.RevisionSpecifier
+	revs           []query.RevisionSpecifier
 }
 
 // given a repo name, determine whether it matched any patterns for which we have
 // revspecs (or ref globs), and if so, return the matching/allowed ones.
-func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []search.RevisionSpecifier, clashing []search.RevisionSpecifier) {
-	revLists := make([][]search.RevisionSpecifier, 0, len(pats))
+func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []query.RevisionSpecifier, clashing []query.RevisionSpecifier) {
+	revLists := make([][]query.RevisionSpecifier, 0, len(pats))
 	for _, rev := range pats {
 		if rev.includePattern.MatchString(string(repo)) {
 			revLists = append(revLists, rev.revs)
@@ -884,14 +884,14 @@ func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []
 	}
 	// no matches: we generate a dummy list containing only master
 	if len(revLists) == 0 {
-		matched = []search.RevisionSpecifier{{RevSpec: ""}}
+		matched = []query.RevisionSpecifier{{RevSpec: ""}}
 		return
 	}
 	// if two repo specs match, and both provided non-empty rev lists,
 	// we want their intersection, so we count the number of times we
 	// see a revision in the rev lists, and make sure it matches the number
 	// of rev lists
-	revCounts := make(map[search.RevisionSpecifier]int, len(revLists[0]))
+	revCounts := make(map[query.RevisionSpecifier]int, len(revLists[0]))
 
 	var aliveCount int
 	for i, revList := range revLists {
@@ -905,7 +905,7 @@ func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []
 	}
 
 	if aliveCount > 0 {
-		matched = make([]search.RevisionSpecifier, 0, len(revCounts))
+		matched = make([]query.RevisionSpecifier, 0, len(revCounts))
 		for rev, seenCount := range revCounts {
 			if seenCount == len(revLists) {
 				matched = append(matched, rev)
@@ -915,7 +915,7 @@ func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []
 		return
 	}
 
-	clashing = make([]search.RevisionSpecifier, 0, len(revCounts))
+	clashing = make([]query.RevisionSpecifier, 0, len(revCounts))
 	for rev := range revCounts {
 		clashing = append(clashing, rev)
 	}
@@ -927,7 +927,7 @@ func getRevsForMatchedRepo(repo api.RepoName, pats []patternRevspec) (matched []
 // findPatternRevs separates out each repo filter into its repository name
 // pattern and its revision specs (if any). It validates each repository name
 // pattern and applies some small optimizations.
-func findPatternRevs(includePatterns []search.ParsedRepoFilter) (outputPatterns []string, includePatternRevs []patternRevspec, err error) {
+func findPatternRevs(includePatterns []query.ParsedRepoFilter) (outputPatterns []string, includePatternRevs []patternRevspec, err error) {
 	outputPatterns = make([]string, 0, len(includePatterns))
 	includePatternRevs = make([]patternRevspec, 0, len(includePatterns))
 
@@ -989,7 +989,7 @@ func (MissingRepoRevsError) Error() string { return "missing repo revs" }
 
 type RepoRevSpecs struct {
 	Repo types.MinimalRepo
-	Revs []search.RevisionSpecifier
+	Revs []query.RevisionSpecifier
 }
 
 func (r *RepoRevSpecs) RevSpecs() []string {
