+++ internal/search/zoekt/indexed_search_test.go
@@ -818,7 +818,7 @@ func TestContextWithoutDeadline_cancel(t *testing.T) {
 func makeRepositoryRevisions(repos ...string) []*search.RepositoryRevisions {
 	r := make([]*search.RepositoryRevisions, len(repos))
 	for i, repospec := range repos {
-		repoRevs := search.ParseRepositoryRevisions(repospec)
+		repoRevs := query.ParseRepositoryRevisions(repospec)
 		revs := make([]string, 0, len(repoRevs.Revs))
 		for _, revSpec := range repoRevs.Revs {
 			revs = append(revs, revSpec.RevSpec)
